move convert_charge_string to excel_helper
